Issue#548: FixedPollingStrategy blocks if remote source is unavailable.
[Problem]

When instantiating a PollingDynamicConfig using the FixedPollingStrategy, the first update to populate the config cache is run on the main thread indefinitely until it succeeds. Therefore, if the config reading (reader::call()) fails the ain thread gets stuck in an infinite loop.

[Solution]

Execute the first update to populate the config cache in the background thread using the ScheduledExecutorService.

[Testing]

./gradlew test

Added unit test to reproduce the issue.